Move to new component
