+++ vue/components/ui/templates/home-part/home-part.stories.js
@@ -1,79 +1,28 @@
 import { storiesOf } from "@storybook/vue";
-import { withKnobs, text, select } from "@storybook/addon-knobs";
+import { withKnobs, text } from "@storybook/addon-knobs";
 
 storiesOf("Templates", module)
     .addDecorator(withKnobs)
     .add("Home", () => ({
         props: {
             logo: {
-                default: text("Logo", require("./logo.svg"))
+                default: text("Logo", "https://cdn.platforme.com/images/platforme.png")
             },
             illustration: {
-                default: text("Illustration", require("./illustration.svg"))
-            },
-            background: {
-                default: text("Background", require("./background.svg"))
+                default: text(
+                    "Illustration",
+                    "https://assets-ouch.icons8.com/preview/230/0a740638-d82c-4ee1-a169-c32096e14e8a.png"
+                )
             },
             message: {
-                default: text("Message", "Welcome to the\nRIPE White Admin")
-            },
-            subMessage: {
-                default: text(
-                    "Sub Message",
-                    "The place where you can create your\ncustomisations, gifts and much more."
-                )
-            },
-            topColor: {
-                default: select(
-                    "Top Color",
-                    {
-                        Dark: "dark",
-                        Light: "light"
-                    },
-                    "dark"
-                )
-            },
-            middleColor: {
-                default: select(
-                    "Middle Color",
-                    {
-                        Dark: "dark",
-                        Light: "light"
-                    },
-                    "dark"
-                )
-            },
-            bottomColor: {
-                default: select(
-                    "Bottom Color",
-                    {
-                        Dark: "dark",
-                        Light: "light"
-                    },
-                    "light"
-                )
-            },
-            links: {
-                default: () => [
-                    { name: "Products", to: "https://www.platforme.com/" },
-                    { name: "Prices", to: { name: "someRoute" } }
-                ]
+                default: text("Message", "Welcome to RIPE XXXX\nyour number one place for RIPE")
             }
         },
         template: `
-            <div>
-                <global />
-                <home-part
-                    v-bind:logo="logo"
-                    v-bind:illustration="illustration"
-                    v-bind:background="background"
-                    v-bind:message="message"
-                    v-bind:sub-message="subMessage"
-                    v-bind:links="links"
-                    v-bind:top-color="topColor"
-                    v-bind:middle-color="middleColor"
-                    v-bind:bottom-color="bottomColor"
-                />
-            </div>
+            <home-part
+                v-bind:logo="logo"
+                v-bind:illustration="illustration"
+                v-bind:message="message"
+            />
         `
     }));
